+++ src/bin/pass-it-on-server.rs
@@ -1,8 +1,8 @@
 use clap::Parser;
 use log::{error, info, LevelFilter};
-use pass_it_on::start_server;
 use pass_it_on::Error;
 use pass_it_on::ServerConfiguration;
+use pass_it_on::{start_server, verify_matrix_devices};
 use std::path::PathBuf;
 
 const LOG_TARGET: &str = "pass_it_on_server";
@@ -16,6 +16,9 @@ struct CliArgs {
     /// Set the logging level [default: Info]
     #[clap(short, long, value_parser)]
     log_level: Option<LevelFilter>,
+    /// Interactively verify Matrix endpoint devices when set
+    #[clap(short, long, value_parser, default_value_t = false)]
+    matrix_verify_devices: bool,
 }
 
 #[tokio::main]
@@ -35,13 +38,14 @@ async fn main() {
         error!(target: LOG_TARGET, "{}", error)
     }
 }
-async fn run(args: CliArgs) -> Result<(), Error> {
+async fn run(cliargs: CliArgs) -> Result<(), Error> {
+    info!(target: LOG_TARGET, "Log level is set to {}", cliargs.log_level.unwrap_or(LevelFilter::Info));
     // Setup default directories
     let default_config_path = directories::ProjectDirs::from("com", "pass-it-on", "pass-it-on-server").unwrap();
 
     // Parse Config file
     let server_config = {
-        let config_path = match args.configuration {
+        let config_path = match cliargs.configuration {
             Some(path) => path,
             None => PathBuf::from(default_config_path.config_dir()).join("server.toml"),
         };
@@ -50,5 +54,9 @@ async fn run(args: CliArgs) -> Result<(), Error> {
         ServerConfiguration::try_from(std::fs::read_to_string(config_path)?.as_str())?
     };
 
-    start_server(server_config, None, None).await
+    // Run interactive matrix device verification when flag is passed
+    match cliargs.matrix_verify_devices {
+        true => verify_matrix_devices(server_config).await,
+        false => start_server(server_config, None, None).await,
+    }
 }