+++ src/lib.rs
@@ -97,6 +97,8 @@ pub use self::configuration::ServerConfiguration;
 pub use self::error::Error;
 #[cfg(feature = "server")]
 pub use self::server::start_server;
+#[cfg(all(feature = "server", feature = "matrix"))]
+pub use self::server::verify_matrix_devices;
 
 /// Logging target value used for the library.
 pub const LIB_LOG_TARGET: &str = "pass_it_on";